feat: customizable eulexia fab
