+++ src/EulexiaFab/EulexiaFab.jsx
@@ -4,12 +4,28 @@ import { EulexiaContextProvider } from '../Context/eulexia-context.jsx'
 
 import Fab from '../Fab/Fab.jsx'
 import RulerSetup from '../EulexiaAction/RulerSetup/RulerSetup.jsx'
+import { ConfigIcon } from '../Icon/index.jsx'
 
-const EulexiaFab = ({ children }) => {
+const EulexiaFab = ({
+  id = 'eulexiaFab',
+  backgroundColor = '#A7C5E6',
+  style = { left: 20, bottom: 20 },
+  event = 'hover',
+  icon = <ConfigIcon />,
+  children
+}) => {
   return (
     <EulexiaContextProvider>
       <RulerSetup />
-      <Fab>{children}</Fab>
+      <Fab
+        id={id}
+        mainButtonStyles={{ backgroundColor }}
+        style={style}
+        icon={icon}
+        event={event}
+      >
+        {children}
+      </Fab>
     </EulexiaContextProvider>
   )
 }
